+++ packages/types/src/interfaces/chain/definitions.ts
@@ -15,7 +15,7 @@ export default {
         {
           name: 'hash',
           type: 'BlockHash',
-          isCached: true,
+          isHistoric: true,
           isOptional: true
         }
       ],